Performance related changes
Going from reading indexes in reader.py to iterating and constructing
dictionaries gave a small bump to execution times.

The bigger difference is in the test data creation, which is now created
an written in batches as opposed getting created in one large object as
before. This makes bulk testing much easier!